deleted components copy folder and created separate component folders for each component
