+++ src/App.js
@@ -4,6 +4,8 @@ import FAQ from './components/Accordion/FAQ';
 import Footer from './components/Footer/Footer';
 import Team from "./components/Team/Team";
 import LoginSignup from "./components/LoginSignup/LoginSignup";
+import { Service } from "./components/Services/Service";
+import { Portfolio } from "./components/Portfolio/Portfolio";
 import './App.css';
 
 
@@ -12,14 +14,14 @@ function App() {
 
   return (
     <div>
-      <BrowserRouter>
+      { <BrowserRouter>
         <Routes>
-          <Route path='/' element={<><LoginSignup /></>} />
-          <Route path='/team' element={<><FAQ /><Team /><Footer /></>} />
+        <Route path='/' element={<><LoginSignup /></>} />
+        <Route path='/team' element={<><FAQ /><Team /><Footer /></>} />
         </Routes>
-      </BrowserRouter>
-
-     
+      </BrowserRouter>     }
+      <Service/>
+      <Portfolio/>
     </div>
   );
 }
