--- src/components copy/Portfolio.jsx
@@ -1,11 +1,0 @@
-import { ImageContainer } from "./ImageContainer";
-
-export function Portfolio() {
-  return (
-    <>
-      <h4 className="portfolio">PORTFOLIO</h4>
-      <h1>WHAT WE'VE DONE</h1>
-      <ImageContainer />
-    </>
-  );
-}
