--- src/components copy/Service.jsx
@@ -1,11 +1,0 @@
-import CardsContainer from "./CardsContainer";
-
-export function Service() {
-  return (
-    <>
-      <h4 className="services">Services</h4>
-      <h1>WHAT WE DO OFFER</h1>
-      <CardsContainer />
-    </>
-  );
-};
