+++ src/components/Portfolio/ImageContainer.css
@@ -1,0 +1,6 @@
+.imageContainer{
+    display: grid;
+    grid-template-columns: repeat(3,1fr);
+    grid-gap: 15px;
+  }
+
