+++ src/components/Portfolio/Portfolio.css
@@ -1,0 +1,7 @@
+.portfolio {
+  font-size: -0.5em;
+  font-family: "Poppins";
+  color: darkgrey;
+  letter-spacing: 2px;
+  text-transform: uppercase;
+}
