+++ src/components/Portfolio/Portfolio.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import { ImageContainer } from "./ImageContainer";
+import "./Portfolio.css"
+
+export function Portfolio() {
+    return (
+        <>
+            <h4 className="portfolio">PORTFOLIO</h4>
+            <h1>WHAT WE'VE DONE</h1>
+            <ImageContainer />
+        </>
+    );
+}
