+++ src/components/Services/Cards.css
@@ -1,0 +1,39 @@
+.card {
+    padding: 2px;
+    width: 363px;
+    height: 209px;
+    border: 2px solid white;
+    border-radius: 10px;
+    margin: 10px;
+    box-shadow: 0 2px 29px 0 rgb(68 88 144 / 16%);
+    transition: all 0.4s ease-in-out;
+    display: flex;
+    align-items: center;
+    font-size: larger;
+  }
+  
+  .card:hover{
+    cursor: pointer;
+    box-shadow: 0 2px 29px 0 rgba(4, 5, 8, 0.256);
+  
+  }
+  
+   .content {
+    margin-left: 15px;
+  }
+  
+  a {
+    text-decoration: none;
+    color: black;
+  }
+  
+  a:hover {
+    color: #fd9042;
+    transition: ease-in-out 0.3s;
+  }
+  
+  .description{
+    font-size: 14px;
+    font-weight: 200;
+  } 
+
