+++ src/components/Services/CardsContainer.css
@@ -1,0 +1,5 @@
+.card-container {
+    display: flex;
+    flex-wrap: wrap;
+  }
+
