+++ src/components/Services/Service.css
@@ -1,0 +1,7 @@
+.services{
+    font-size: -0.5em;
+    font-family: "Poppins";
+    color: darkgrey;
+    letter-spacing: 2px;
+    text-transform: uppercase; 
+}
