+++ src/components/Services/Service.js
@@ -1,0 +1,13 @@
+import React from "react";
+import CardsContainer from "./CardsContainer";
+import "./Service.css"
+
+export function Service() {
+    return (
+        <>
+            <h4 className="services">Services</h4>
+            <h1>WHAT WE DO OFFER</h1>
+            <CardsContainer />
+        </>
+    );
+};
